// src/sdk/generated/models/ApiError.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApiError
 */
export interface ApiError {
    /**
     * Machine-readable internal error code.
     * @type {string}
     * @memberof ApiError
     */
    code: string;
    /**
     * 
     * @type {string}
     * @memberof ApiError
     */
    message: string;
    /**
     * 
     * @type {number}
     * @memberof ApiError
     */
    status: number;
    /**
     * 
     * @type {string}
     * @memberof ApiError
     */
    traceId?: string;
    /**
     * 
     * @type {object}
     * @memberof ApiError
     */
    details?: object;
}

/**
 * Check if a given object implements the ApiError interface.
 */
export function instanceOfApiError(value: object): value is ApiError {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    return true;
}

export function ApiErrorFromJSON(json: any): ApiError {
    return ApiErrorFromJSONTyped(json, false);
}

export function ApiErrorFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiError {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'],
        'message': json['message'],
        'status': json['status'],
        'traceId': json['traceId'] == null ? undefined : json['traceId'],
        'details': json['details'] == null ? undefined : json['details'],
    };
}

export function ApiErrorToJSON(json: any): ApiError {
    return ApiErrorToJSONTyped(json, false);
}

export function ApiErrorToJSONTyped(value?: ApiError | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'message': value['message'],
        'status': value['status'],
        'traceId': value['traceId'],
        'details': value['details'],
    };
}